Created tests for PESEL validatation in user serializer

+++ library/books/models/user.py
@@ -3,7 +3,7 @@ from django.contrib.auth.models import AbstractUser
 
 
 class Reader(AbstractUser):
-    PESEL = models.IntegerField(blank=True, null=True)
+    PESEL = models.BigIntegerField(blank=True, null=True)
     telephone = models.IntegerField(blank=True, null=True)
     birth_date = models.DateField(blank=True, null=True)
     address = models.CharField(max_length=200, blank=True, null=True)
